abstract track_content route logic into a component + update tests
