+++ creator-node/test/nodesync.test.js
@@ -4,6 +4,7 @@ const path = require('path')
 const assert = require('assert')
 const _ = require('lodash')
 const nock = require('nock')
+const sinon = require('sinon')
 
 const config = require('../src/config')
 const models = require('../src/models')
@@ -16,13 +17,12 @@ const ipfsImport = require('../src/ipfsClient')
 // TODO - upgrade to newer ipfs client
 const ipfsClient = ipfsImport.ipfs
 const redisClient = require('../src/redis')
-const { wait, stringifiedDateFields } = require('./lib/utils')
+const { stringifiedDateFields } = require('./lib/utils')
+const { uploadTrack } = require('./lib/helpers')
 
 const testAudioFilePath = path.resolve(__dirname, 'testTrack.mp3')
 const sampleExportPath = path.resolve(__dirname, 'syncAssets/sampleExport.json')
 const sampleExportFromClock2Path = path.resolve(__dirname, 'syncAssets/sampleExportFromClock2.json')
-
-const MAX_TRANSCODE_TRACK_TIMEOUT = 300000 // 5 min
 
 describe('test nodesync', async function () {
   let server, app
@@ -49,6 +49,7 @@ describe('test nodesync', async function () {
    * Wipe DB, server, and redis state
    */
   afterEach(async function () {
+    await sinon.restore()
     await server.close()
   })
 
@@ -90,39 +91,7 @@ describe('test nodesync', async function () {
 
       /** Upload a track */
 
-      const file = fs.readFileSync(testAudioFilePath)
-
-      // Upload track content
-      const { body: { data: { uuid } } } = await request(app)
-        .post('/track_content')
-        .attach('file', file, { filename: 'fname.mp3' })
-        .set('Content-Type', 'multipart/form-data')
-        .set('X-Session-ID', sessionToken)
-
-      const start = Date.now()
-      let trackUploadResponse
-      while (Date.now() - start < MAX_TRANSCODE_TRACK_TIMEOUT) {
-        const { body: { data: { status, resp } } } = await request(app)
-          .get('/processing_status')
-          .query({
-            taskType: 'transcode',
-            uuid: uuid
-          })
-
-        // Should have a body structure of:
-        //   { transcodedTrackCID, transcodedTrackUUID, track_segments, source_file }
-        if (status && status === 'DONE') {
-          trackUploadResponse = resp.object.data
-          break
-        }
-
-        if (status && status === 'FAILED') throw new Error(`Transcode failed: ${JSON.stringify(resp)}`)
-
-        // Check the transcode status every 1s
-        await wait(1000)
-      }
-
-      if (!trackUploadResponse) throw new Error(`Transcode exceeded max timeout=${MAX_TRANSCODE_TRACK_TIMEOUT}ms`)
+      const trackUploadResponse = (await uploadTrack(testAudioFilePath, cnodeUserUUID)).data
 
       transcodedTrackUUID = trackUploadResponse.transcodedTrackUUID
       trackSegments = trackUploadResponse.track_segments
